Create reusable opening period accordion component
